+++ tcex/testing/test_case_job.py
@@ -1,12 +1,15 @@
 # -*- coding: utf-8 -*-
 """TcEx Runtime App Test Case"""
 import os
-
+import subprocess
+import sys
 from .test_case import TestCase
 
 
 class TestCaseJob(TestCase):
     """App TestCase Class"""
+
+    run_method = 'inline'  # run service inline or a as subprocess
 
     @staticmethod
     def create_shelf_dir(shelf_path):
@@ -16,47 +19,48 @@ class TestCaseJob(TestCase):
             with open(os.path.join(shelf_path, 'DEBUG'), 'a'):
                 os.utime(os.path.join(shelf_path, 'DEBUG'), None)
 
-    def run(self, args):  # pylint: disable=too-many-return-statements
+    def run(self):
         """Run the Playbook App.
-
-        Args:
-            args (dict): The App CLI args.
 
         Returns:
             int: The exit code fo the App.
         """
-        # safely log all args to tests.log
-        self._log_args(args)
+        from run import run
 
-        # get a configured instance of the App
-        self.app = self.app_init(args)
-
-        # Setup
-        exit_code = self.run_app_method(self.app, 'setup')
-        if exit_code != 0:
-            return exit_code
-
-        # Run
-        exit_code = self.run_app_method(self.app, 'run')
-        if exit_code != 0:
-            return exit_code
-
-        # Teardown
-        exit_code = self.run_app_method(self.app, 'teardown')
-        if exit_code != 0:
-            return exit_code
-
+        # run the app
+        exit_code = 0
         try:
-            # call exit for message_tc output, but don't exit
-            self.app.tcex.playbook.exit(msg=self.app.exit_message)
-        except SystemExit:
-            pass
+            run()
+        except SystemExit as e:
+            exit_code = e.code
 
-        return self._exit(self.app.tcex.exit_code)
+        self.log.data('run', 'Exit Code', exit_code)
+        return exit_code
 
     def run_profile(self):
         """Run an App using the profile name."""
         self.create_shelf_dir(self.profile.tc_temp_path)
 
+        # create encrypted config file
+        self.create_config(self.profile.args)
+
+        # run the service App in 1 of 3 ways
+        if self.run_method == 'inline':
+            # backup sys.argv
+            sys_argv_orig = sys.argv
+
+            # clear sys.argv
+            sys.argv = sys.argv[:1]
+
+            # run the App
+            exit_code = self.run()
+
+            # restore sys.argv
+            sys.argv = sys_argv_orig
+        elif self.run_method == 'subprocess':
+            # run the Service App as a subprocess
+            app_process = subprocess.Popen(['python', 'run.py'])
+            exit_code = app_process.wait()
+
         # run the App
-        return self.run(self.profile.args)
+        return exit_code