+++ tcex/testing/test_case_trigger_service.py
@@ -1,65 +1,36 @@
 # -*- coding: utf-8 -*-
 """TcEx Service Common Module"""
-import traceback
-
+import sys
 from .test_case_service_common import TestCaseServiceCommon
 
 
 class TestCaseTriggerService(TestCaseServiceCommon):
     """Service App TestCase Class"""
 
-    def run(self, args):  # pylint: disable=too-many-return-statements
+    # TOOD: move to service command
+    def run(self):
         """Run the Playbook App.
-
-        Args:
-            args (dict): The App CLI args.
 
         Returns:
             int: The App exit code
         """
-        # safely log all args to tests.log
-        self._log_args(args)
+        from run import run
 
-        # get a configured instance of the App
-        self.app = self.app_init(args)
+        # backup sys.argv
+        sys_argv_orig = sys.argv
 
-        # Setup
-        exit_code = self.run_app_method(self.app, 'setup')
-        if exit_code != 0:
-            return exit_code
+        # clear sys.argv
+        sys.argv = sys.argv[:1]
 
-        # Trigger
+        # run the app
+        exit_code = 0
         try:
-            # configure custom trigger message handler
-            self.app.tcex.service.create_config_callback = self.app.create_config_callback
-            self.app.tcex.service.delete_config_callback = self.app.delete_config_callback
-            self.app.tcex.service.shutdown_callback = self.app.shutdown_callback
-
-            self.app.tcex.service.listen()
-            self.app.tcex.service.heartbeat()
-            self.app.tcex.service.ready = True
+            run()
         except SystemExit as e:
-            self.log.data(
-                'run', 'App failed', f'App exited with code of {e.code} in method run()', 'error',
-            )
-            return self._exit(e.code)
-        except Exception:
-            self.log.data(
-                'run',
-                'App failed',
-                f'App encountered except in run() method ({traceback.format_exc()})',
-                'error',
-            )
-            return self._exit(1)
+            exit_code = e.code
 
-        # Run
-        exit_code = self.run_app_method(self.app, 'run')
-        if exit_code != 0:
-            return exit_code
+        # restore sys.argv
+        sys.argv = sys_argv_orig
 
-        # Teardown
-        exit_code = self.run_app_method(self.app, 'teardown')
-        if exit_code != 0:
-            return exit_code
-
-        return self._exit(self.app.tcex.exit_code)
+        self.log.data('run', 'Exit Code', exit_code)
+        return exit_code