+++ docs/src/release_notes.rst
@@ -3,6 +3,12 @@
 Release Notes
 #############
 
+2.0.5
+=====
++ Updated testing framework to decouple App version of TcEx and testing version.
++ Updated deepdiff validation method to better handle OrderedDicts.
++ Added simple caching to env_store.
+
 2.0.4
 =====
 + Updated decorator method logging.